amend deps, fix react types, organise ec plugins
