+++ package.json
@@ -17,7 +17,7 @@
 		"format:core:check": "npm run format:core -- --check",
 		"format:content": "npx prettier --write \"**/*.{md,mdx,astro}\"",
 		"format:data": "npx prettier --write \"**/*.{json,yaml,yml}\"",
-		"postinstall": "npx patch-package && npm run sync",
+		"postinstall": "npm run sync",
 		"preview": "npx astro preview",
 		"script:optimize-svgs": "npx tsx scripts/optimize-svgs.ts",
 		"start": "npx astro dev",
@@ -26,85 +26,85 @@
 		"lint": "npx eslint"
 	},
 	"devDependencies": {
-		"@astro-community/astro-embed-youtube": "^0.5.6",
-		"@astrojs/check": "^0.9.4",
-		"@astrojs/react": "^3.6.3",
-		"@astrojs/rss": "^4.0.11",
-		"@astrojs/sitemap": "^3.2.1",
-		"@astrojs/starlight": "^0.29.3",
-		"@astrojs/starlight-docsearch": "^0.3.0",
-		"@astrojs/starlight-tailwind": "^2.0.3",
-		"@astrojs/tailwind": "^5.1.4",
-		"@cloudflare/puppeteer": "^0.0.14",
-		"@cloudflare/vitest-pool-workers": "^0.6.0",
-		"@cloudflare/workers-types": "^4.20241218.0",
-		"@codingheads/sticky-header": "^1.0.2",
-		"@expressive-code/plugin-collapsible-sections": "^0.40.0",
-		"@iarna/toml": "^2.2.5",
-		"@iconify-json/material-symbols": "^1.2.12",
-		"@stoplight/json-schema-tree": "^4.0.0",
-		"@types/hast": "^3.0.4",
-		"@types/he": "^1.2.3",
-		"@types/node": "^22.10.6",
-		"@types/react": "^18.3.12",
-		"@types/react-dom": "^18.3.1",
-		"@typescript-eslint/parser": "^8.20.0",
-		"algoliasearch": "^5.19.0",
-		"astro": "^4.16.18",
-		"astro-breadcrumbs": "^3.3.1",
-		"astro-icon": "^1.1.2",
-		"astro-live-code": "^0.0.5",
-		"date-fns": "^4.1.0",
-		"dedent": "^1.5.3",
-		"detype": "1.0.12",
+		"@astrojs/check": "0.9.4",
+		"@astrojs/react": "4.1.5",
+		"@astrojs/rss": "4.0.11",
+		"@astrojs/sitemap": "3.2.1",
+		"@astrojs/starlight": "0.29.3",
+		"@astrojs/starlight-docsearch": "0.3.0",
+		"@astrojs/starlight-tailwind": "2.0.3",
+		"@astrojs/tailwind": "5.1.4",
+		"@cloudflare/puppeteer": "0.0.14",
+		"@cloudflare/vitest-pool-workers": "0.6.0",
+		"@cloudflare/workers-types": "4.20250109.0",
+		"@codingheads/sticky-header": "1.0.2",
+		"@expressive-code/plugin-collapsible-sections": "0.38.3",
+		"@iarna/toml": "2.2.5",
+		"@iconify-json/material-symbols": "1.2.12",
+		"@stoplight/json-schema-tree": "4.0.0",
+		"@types/hast": "3.0.4",
+		"@types/he": "1.2.3",
+		"@types/node": "22.10.7",
+		"@types/react": "19.0.7",
+		"@types/react-dom": "19.0.3",
+		"@typescript-eslint/parser": "8.20.0",
+		"algoliasearch": "5.19.0",
+		"astro": "4.16.18",
+		"astro-breadcrumbs": "3.3.1",
+		"astro-icon": "1.1.5",
+		"astro-live-code": "0.0.5",
+		"date-fns": "4.1.0",
+		"dedent": "1.5.3",
 		"dompurify": "3.2.3",
-		"dot-prop": "^9.0.0",
-		"eslint": "^9.18.0",
-		"eslint-plugin-astro": "^1.3.1",
-		"eslint-plugin-jsx-a11y": "^6.10.2",
-		"eslint-plugin-react": "^7.37.4",
-		"fast-glob": "^3.3.3",
-		"github-slugger": "^2.0.0",
-		"globals": "^15.14.0",
-		"hastscript": "^9.0.0",
-		"he": "^1.2.0",
-		"instantsearch.css": "^8.5.1",
-		"instantsearch.js": "^4.75.7",
-		"jsonc-parser": "^3.3.1",
-		"lz-string": "^1.5.0",
-		"marked": "^15.0.6",
-		"mdast-util-mdx-expression": "^2.0.1",
-		"mermaid": "^11.4.1",
-		"node-html-parser": "^6.1.13",
-		"patch-package": "^8.0.0",
-		"playwright": "^1.49.1",
-		"prettier": "^3.4.2",
-		"prettier-plugin-astro": "^0.14.1",
-		"prettier-plugin-tailwindcss": "^0.6.9",
-		"puppeteer": "^24.0.0",
-		"react": "^18.3.1",
-		"react-dom": "^18.3.1",
-		"react-markdown": "^9.0.3",
-		"redirects-in-workers": "^0.0.5",
-		"rehype-autolink-headings": "^7.1.0",
-		"rehype-external-links": "^3.0.0",
-		"rehype-mermaid": "^3.0.0",
-		"rehype-title-figure": "^0.1.2",
-		"sharp": "^0.33.5",
-		"solarflare-theme": "^0.0.2",
-		"starlight-image-zoom": "^0.9.0",
-		"starlight-links-validator": "^0.14.1",
-		"starlight-package-managers": "^0.9.0",
-		"svgo": "^3.3.2",
-		"tailwindcss": "^3.4.17",
-		"tippy.js": "^6.3.7",
-		"tsx": "^4.19.2",
-		"typescript": "^5.7.3",
-		"typescript-eslint": "^8.20.0",
-		"unist-util-visit": "^5.0.0",
+		"dot-prop": "9.0.0",
+		"eslint": "9.18.0",
+		"eslint-plugin-astro": "1.3.1",
+		"eslint-plugin-jsx-a11y": "6.10.2",
+		"eslint-plugin-react": "7.37.4",
+		"fast-glob": "3.3.3",
+		"github-slugger": "2.0.0",
+		"globals": "15.14.0",
+		"hastscript": "9.0.0",
+		"he": "1.2.0",
+		"instantsearch.css": "8.5.1",
+		"instantsearch.js": "4.77.0",
+		"jsonc-parser": "3.3.1",
+		"lz-string": "1.5.0",
+		"marked": "15.0.6",
+		"mdast-util-mdx-expression": "2.0.1",
+		"mermaid": "11.4.1",
+		"node-html-parser": "7.0.1",
+		"playwright": "1.49.1",
+		"prettier": "3.4.2",
+		"prettier-plugin-astro": "0.14.1",
+		"prettier-plugin-tailwindcss": "0.6.9",
+		"puppeteer": "24.1.0",
+		"react": "19.0.0",
+		"react-dom": "19.0.0",
+		"react-markdown": "9.0.3",
+		"redirects-in-workers": "0.0.5",
+		"rehype-autolink-headings": "7.1.0",
+		"rehype-external-links": "3.0.0",
+		"rehype-mermaid": "3.0.0",
+		"rehype-title-figure": "0.1.2",
+		"sharp": "0.33.5",
+		"solarflare-theme": "0.0.2",
+		"starlight-image-zoom": "0.9.0",
+		"starlight-links-validator": "0.14.1",
+		"starlight-package-managers": "0.9.0",
+		"svgo": "3.3.2",
+		"tailwindcss": "3.4.17",
+		"tippy.js": "6.3.7",
+		"ts-blank-space": "0.5.0",
+		"tsx": "4.19.2",
+		"typescript": "5.7.3",
+		"typescript-eslint": "8.20.0",
+		"unist-util-visit": "5.0.0",
 		"vitest": "2.1.8",
-		"wrangler": "^3.101.0",
-		"yaml": "^2.7.0"
+		"wrangler": "3.101.0"
+	},
+	"optionalDependencies": {
+		"@rollup/rollup-linux-x64-gnu": "*"
 	},
 	"engines": {
 		"node": ">=22"
