+++ plugins/expressive-code/default-titles.js
@@ -1,0 +1,20 @@
+import { definePlugin } from "@expressive-code/core";
+
+export default () => {
+	return definePlugin({
+		name: "Adds language-specific default titles.",
+		hooks: {
+			preprocessLanguage: async (context) => {
+				switch (context.codeBlock.language) {
+					case "powershell": {
+						context.codeBlock.props.title ??= "PowerShell";
+						break;
+					}
+					default: {
+						return;
+					}
+				}
+			},
+		},
+	});
+};
