+++ src/pages/ruleset-engine/rules-language/fields/reference/index.astro
@@ -1,7 +1,7 @@
 ---
 import { getEntry } from "astro:content";
 import StarlightPage from "@astrojs/starlight/components/StarlightPage.astro";
-import FieldCatalog from "~/components/FieldCatalog.jsx";
+import FieldCatalog from "~/components/FieldCatalog.tsx";
 
 const fieldData = await getEntry("fields", "index");
 const fields = fieldData.data.entries;
