+++ src/pages/workers-ai/models/[name].astro
@@ -5,8 +5,8 @@ import StarlightPage, {
 	type StarlightPageProps,
 } from "@astrojs/starlight/components/StarlightPage.astro";
 import { LinkButton, Tabs, TabItem, Code, Aside } from "~/components";
-import ModelInfo from "~/components/models/ModelInfo.jsx";
-import ModelBadges from "~/components/models/ModelBadges.jsx";
+import ModelInfo from "~/components/models/ModelInfo.tsx";
+import ModelBadges from "~/components/models/ModelBadges.tsx";
 import SchemaViewer from "~/components/models/SchemaViewer.astro";
 
 import TextGenerationCode from "~/components/models/code/TextGenerationCode.astro";
